refactor: 모든 상품 조회 API 재구현 및 상품 테스트 픽스쳐 패키지 위치 수정
