+++ src/main/java/kitchenpos/products/ui/response/ProductResponse.java
@@ -1,6 +1,9 @@
 package kitchenpos.products.ui.response;
 
+import static java.util.stream.Collectors.*;
+
 import java.math.BigDecimal;
+import java.util.List;
 import java.util.UUID;
 import kitchenpos.products.domain.Product;
 
@@ -22,6 +25,12 @@ public class ProductResponse {
             product.getNameValue(),
             product.getPriceValue()
         );
+    }
+
+    public static List<ProductResponse> of(List<Product> products) {
+        return products.stream()
+            .map(ProductResponse::from)
+            .collect(toList());
     }
 
     public UUID getId() {
