+++ src/test/java/kitchenpos/eatinorders/application/OrderTableServiceTest.java
@@ -14,8 +14,8 @@ import org.junit.jupiter.params.provider.ValueSource;
 import java.util.List;
 import java.util.UUID;
 
-import static kitchenpos.asis.Fixtures.order;
-import static kitchenpos.asis.Fixtures.orderTable;
+import static kitchenpos.Fixtures.order;
+import static kitchenpos.Fixtures.orderTable;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
