+++ src/test/java/kitchenpos/products/ProductFixtures.java
@@ -1,4 +1,4 @@
-package kitchenpos.tobe;
+package kitchenpos.products;
 
 import java.math.BigDecimal;
 import java.util.UUID;
@@ -7,7 +7,7 @@ import kitchenpos.products.domain.Product;
 import kitchenpos.products.domain.ProductName;
 import kitchenpos.products.domain.ProductPrice;
 
-public class Fixtures {
+public class ProductFixtures {
 
     public static final UUID INVALID_ID = new UUID(0, 0);
