+++ src/test/java/kitchenpos/products/domain/ProductTest.java
@@ -1,9 +1,9 @@
 package kitchenpos.products.domain;
 
+import static kitchenpos.products.ProductFixtures.product;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.UUID;
-import kitchenpos.tobe.Fixtures;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -14,12 +14,12 @@ class ProductTest {
     void sameUUID() {
         // given
         UUID id = UUID.fromString("38400000-8cf0-11bd-b23e-10b96e4ef00d");
-        Product apple = Fixtures.product(
+        Product apple = product(
             id,
             "사과",
             3_000
         );
-        Product banana = Fixtures.product(
+        Product banana = product(
             id,
             "바나나",
             5_000
